[Semsterarbeit] javafx
