+++ src/Main/GA_Trial/FXML-View.fxml
@@ -1,0 +1,103 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.geometry.Insets?>
+<?import javafx.collections.FXCollections?>
+<?import javafx.scene.control.cell.PropertyValueFactory?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.image.Image?>
+<GridPane xmlns="http://javafx.com/javafx"
+          xmlns:fx="http://javafx.com/fxml"
+          fx:controller="GA_Trial.FXML_Controller"
+          prefHeight="400.0" prefWidth="600.0"
+          hgap="10" vgap="10" fx:id="gp">
+
+
+
+    <padding>
+        <Insets topRightBottomLeft="20"/>
+    </padding>
+
+    <Label fx:id="erfassung" text="Erfassen" GridPane.columnIndex="0" GridPane.rowIndex="0"
+           GridPane.columnSpan="2" GridPane.rowSpan="1"/>
+    <Label fx:id="standOrt" text="Von" GridPane.columnIndex="0" GridPane.rowIndex="1"/>
+    <Label fx:id="reiseziel" text="Nach" GridPane.columnIndex="1" GridPane.rowIndex="1"/>
+    <Label fx:id="reisedatum" text="Datum" GridPane.columnIndex="2" GridPane.rowIndex="1"/>
+    <Label fx:id="koseten" text="Kosten" GridPane.columnIndex="3" GridPane.rowIndex="1"/>
+    <Label fx:id="chur" text="Chur" GridPane.columnIndex="0" GridPane.rowIndex="2"/>
+    <Label fx:id="auswertung" text="Auswertung" GridPane.columnIndex="0" GridPane.rowIndex="5" GridPane.rowSpan="1" GridPane.columnSpan="2"/>
+    <Label fx:id="preisSum" text="Total Reisekosten" GridPane.columnIndex="2" GridPane.rowIndex="19"/>
+    <Label fx:id="gaRel" text="Preisverhältnis zu GA" GridPane.columnIndex="3" GridPane.rowIndex="19"/>
+
+    <TextField fx:id="price" GridPane.columnIndex="3" GridPane.rowIndex="2"/>
+    <TextField fx:id="kostenTotal" GridPane.columnIndex="2" GridPane.rowIndex="20" GridPane.columnSpan="1" GridPane.rowSpan="2"/>
+    <TextField fx:id="relation" GridPane.columnIndex="3" GridPane.rowIndex="20" GridPane.columnSpan="1" GridPane.rowSpan="2"/>
+    <TextField fx:id="other" GridPane.columnIndex="1" GridPane.rowIndex="3" visible="false"/>
+
+
+
+    <ComboBox fx:id="comboBoxZiel" GridPane.columnIndex="1" GridPane.rowIndex="2" promptText="Bitte auswählen" onAction="#setDestination">
+        <items>
+            <FXCollections fx:factory="observableArrayList">
+                <String fx:value="Zürich" fx:id="Zürich"/>
+                <String fx:value="Bern" fx:id="Bern"/>
+                <String fx:value="Olten" fx:id="Olten"/>
+                <String fx:value="Anderer Zielort" fx:id="others"/>
+            </FXCollections>
+        </items>
+    </ComboBox>
+
+    <ComboBox fx:id="comboBoxMonat" GridPane.columnIndex="0" GridPane.rowIndex="8" GridPane.columnSpan="2" GridPane.rowSpan="1" promptText="Bitte auswählen" onAction="#setMonth">
+        <items>
+            <FXCollections fx:factory="observableArrayList">
+                <String fx:value="Januar"/>
+                <String fx:value="Februar"/>
+                <String fx:value="März"/>
+                <String fx:value="April"/>
+                <String fx:value="Mai"/>
+                <String fx:value="Juni"/>
+                <String fx:value="Juli"/>
+                <String fx:value="August"/>
+                <String fx:value="September"/>
+                <String fx:value="Oktober"/>
+                <String fx:value="November"/>
+                <String fx:value="Dezember"/>
+            </FXCollections>
+        </items>
+    </ComboBox>
+
+    <Button fx:id="enterReise" text="Erfasse Reise" GridPane.columnIndex="3" GridPane.rowIndex="3" alignment="TOP_RIGHT" onAction="#setTrip"/>
+    <Button fx:id="showAll" text="Zeige alle Reisen" GridPane.columnIndex="0" GridPane.rowIndex="9" GridPane.columnSpan="3" GridPane.rowSpan="1" onAction="#showAllTrips"/>
+    <Button fx:id="deleteTrip" text="Reise löschen" GridPane.columnIndex="0" GridPane.rowIndex="10" GridPane.columnSpan="2" GridPane.rowSpan="1" onAction="#deleteTrip"/>
+
+    <DatePicker fx:id="datePicker" GridPane.columnIndex="2" GridPane.rowIndex="2" onAction="#setDatepicker"/>
+
+
+    <BorderPane>
+    </BorderPane>
+
+
+    <TableView fx:id="reiseTable" GridPane.columnIndex="2" GridPane.rowIndex="7" GridPane.columnSpan="2" GridPane.rowSpan="12" minWidth="400">
+        <columns>
+            <TableColumn fx:id="nr" text="ID" visible="false">
+                <cellValueFactory><PropertyValueFactory property="nr"/>
+                </cellValueFactory>
+            </TableColumn>
+            <TableColumn fx:id="destination" text="Destination" minWidth="100">
+                <cellValueFactory><PropertyValueFactory property="destination" />
+                </cellValueFactory>
+            </TableColumn>
+            <TableColumn fx:id="preis" text="Reisekosten" minWidth="100">
+            <cellValueFactory><PropertyValueFactory property="preis"/>
+            </cellValueFactory>
+        </TableColumn>
+            <TableColumn fx:id="datum" text="Reisedatum" minWidth="100">
+                <cellValueFactory><PropertyValueFactory property="date"/>
+                </cellValueFactory>
+            </TableColumn>
+        </columns>
+    </TableView>
+
+</GridPane>
