+++ src/Main/GA_Trial/chart-style.css
@@ -18,3 +18,13 @@
     -fx-background-color: red, white;
 }
 
+
+#auswertung {
+    -fx-font-family: Courier;
+    -fx-font-size: 25px;
+}
+
+#erfassung {
+    -fx-font-family: Courier;
+    -fx-font-size: 25px;
+}
