+++ src/Test/GA_Trial/GUI_Test.java
@@ -1,4 +1,10 @@
 package GA_Trial;
 
-public class GUI_Test {
-}
+public class GUI_Test  {
+
+
+
+
+    }
+
+
