+++ src/Test/GA_Trial/SQL_Persistence_Test.java
@@ -9,11 +9,11 @@ import org.junit.Test;
 import static org.junit.Assert.*;
 
 
-public class SQL_Persistence_Test  {
+public class SQL_Persistence_Test {
 
 
-    @Mock private Connection mockConnection;
-
+    @Mock
+    private Connection mockConnection;
 
     @Before
     public void setUp() {
@@ -23,18 +23,18 @@ public class SQL_Persistence_Test  {
 
 
     @Test
-    public void testSetDatum(){
+    public void testSetDatum() {
         String datum = "19. Dez 2019";
-        Reise reise = new Reise(1, "Bern", Double.parseDouble("100"), LocalDate.of(2019,12, 19));
-        reise.setDatum(LocalDate.of(2019,12,19));
+        Reise reise = new Reise(1, "Bern", Double.parseDouble("100"), LocalDate.of(2019, 12, 19));
+        reise.setDatum(LocalDate.of(2019, 12, 19));
         String result = reise.getDate();
         assertEquals(datum, result);
     }
 
 
     @Test
-    public void shouldTestDBConnectionSQLException(){
-
+    public void shouldTestDBConnectionSQLException() {
+        //Test falls falsche Userdaten eingegeben urden
         try {
             new SQL_Persistence().DBConnection();
             Class.forName("com.mysql.jdbc.Driver");
@@ -48,7 +48,8 @@ public class SQL_Persistence_Test  {
     }
 
     @Test
-    public void shouldTestDBConnectionClassNotFoundException(){
+    public void shouldTestDBConnectionClassNotFoundException() {
+        // Test falls Datenbank-URL beim Verbinungsaufbar nicht korrekt ist.
 
         try {
             new SQL_Persistence().DBConnection();
@@ -63,23 +64,6 @@ public class SQL_Persistence_Test  {
     }
 
 
-    @Test(expected = SQLException.class)
-    public void shouldTestGetTripSQLException(){
-
-        try {
-            new SQL_Persistence().getTrip();
-            Class.forName("com.mysql.jdbc.Driver");
-            mockConnection = DriverManager.getConnection("jdbc:mysql://localhost:3306/calculator", "hallo", "hallo");
-        } catch (ClassNotFoundException ex) {
-            System.out.println("Datenbank nicht gefunden");
-
-        } catch (SQLException ex) {
-            System.out.println("Ein Fehler ist aufgetreten");
-        }
-    }
-
-
-
     @Test
     public void shouldTestCloseConnection() {
         try {
@@ -91,31 +75,8 @@ public class SQL_Persistence_Test  {
 
         } catch (SQLException ex) {
             System.out.println("Ein Fehler ist aufgetreten");
-        }finally {
-            if (mockConnection != null){
-                try {
-                    mockConnection.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-
-
-    @Test
-    public void shouldTestDeleteTripException() {
-        try {
-            new SQL_Persistence().deleteTrip(0);
-            Class.forName("com.mysql.jdbc.Driver");
-            mockConnection = DriverManager.getConnection("jdbc:mysql://localhost:3306/calculator", "hallo", "hallo");
-        } catch (ClassNotFoundException ex) {
-            System.out.println("Datenbank nicht gefunden");
-
-        } catch (SQLException ex) {
-            System.out.println("Ein Fehler ist aufgetreten");
-        }finally {
-            if (mockConnection != null){
+        } finally {
+            if (mockConnection != null) {
                 try {
                     mockConnection.close();
                 } catch (SQLException e) {
@@ -125,3 +86,4 @@ public class SQL_Persistence_Test  {
         }
     }
 }
+
